requirements.txt
streamlit
tensorflow
opencv-python
pillow
numpy
matplotlib
reportlab